commit with 0.0.37 by manju

+++ AuroProctoringSDK/src/main/java/com/example/auroproctoringsdk/utils/Utils.kt
@@ -30,10 +30,12 @@ class Utils() {
         }
         val fileName = "image_${Calendar.getInstance().time}.jpg"
         val file = File(directory, fileName)
-        val outputStream = FileOutputStream(file)
-        image.compress(Bitmap.CompressFormat.JPEG, 100, outputStream)
-        outputStream.flush()
-        outputStream.close()
+        if(file.exists()){
+            val outputStream = FileOutputStream(file)
+            image.compress(Bitmap.CompressFormat.JPEG, 100, outputStream)
+            outputStream.flush()
+            outputStream.close()
+        }
 
         return file.absolutePath
     }
